requirements.txt
Django==4.1.1
django-cors-headers==3.13.0
django-filter==22.1
django-perms==2.1.0
django-templated-mail==1.1.1

djangorestframework

djangorestframework-simplejwt

djoser==2.1.0
Pillow==9.2.0
pipenv==2022.9.24
pipfile==0.0.2
psycopg2==2.9.3
PyJWT==2.5.0
requests==2.28.1
requests-oauthlib==1.3.1
